requirements.txt
appdirs==1.4.0
Django==1.10.5
django-material==0.12.5
django-push-notifications==1.4.1
djangorestframework==3.5.4
gunicorn==19.6.0
linecache2==1.0.0
model-mommy==1.3.2
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
traceback2==1.4.0
unittest2==1.1.0
